tolid-api/requirements.txt
Cython == 0.29.21
connexion == 2.7.0
python_dateutil == 2.8.1
setuptools == 53.0.0
Werkzeug == 1.0.1
Flask-SQLAlchemy == 2.4.4
psycopg2-binary == 2.8.6
openpyxl == 3.0.6
responses
jwt
biopython==1.73
